input_state: Learn fenced floats.
When turned on floats can't be moved or resized so that part of its
window extends outside its output.

When *desktop* mode is used a view can still be moved *between* outputs
unabstructed (or at least that's the plan when sway gets support for
that).

This setting can be set globally and per view.